Refatoração classe Deck
Separado a classe Deck em uma super classe
