+++ src/yugioh/board/player/Field.java
@@ -257,24 +257,6 @@ public class Field {
 
 	}
 
-//	public void addCardToHand() {
-//
-//		if (deck.getDeck().size() == 0) {
-//
-//			if (this == Card.getBoard().getActivePlayer().getField())
-//				Card.getBoard().setWinner(Card.getBoard().getOpponentPlayer());
-//			else
-//				Card.getBoard().setWinner(Card.getBoard().getActivePlayer());
-//
-//			return;
-//		}
-//
-//		Card temp = deck.drawOneCard();
-//		hand.add(temp);
-//		temp.setLocation(Location.HAND);
-//
-//	}
-
 	public void addNCardsToHand(int n) {
 		for (int j = 0; j < n; j++) {
 			if (deck.getDeck().size() == 0) {
